requirements.txt
streamlit>=1.37
pandas>=2.1
numpy>=1.26
python-docx>=1.1
pdfplumber>=0.11
PyYAML>=6.0
xlsxwriter>=3.1
openpyxl>=3.1
python-dateutil>=2.9
